Move router contract to its own file.
